Create ApacheChart component
